+++ surfalize/profile.py
@@ -5,12 +5,12 @@ from scipy.signal import find_peaks
 class Profile:
 
     def __init__(self, height_data, step, length_um):
-        self._data = height_data
-        self._step = step
-        self._length_um = length_um
+        self.data = height_data
+        self.step = step
+        self.length = length_um
 
     def __repr__(self):
-        return f'{self.__class__.__name__}({self._length_um:.2f} µm)'
+        return f'{self.__class__.__name__}({self.length:.2f} µm)'
 
     def _repr_png_(self):
         """
@@ -20,8 +20,8 @@ class Profile:
         self.show()
 
     def period(self):
-        fft = np.abs(np.fft.fft(self._data))
-        freq = np.fft.fftfreq(self._data.shape[0], d=self._step)
+        fft = np.abs(np.fft.fft(self.data))
+        freq = np.fft.fftfreq(self.data.shape[0], d=self.step)
         peaks, properties = find_peaks(fft.flatten(), distance=10, prominence=10)
         # Find the prominence of the peaks
         prominences = properties['prominences']
@@ -35,30 +35,30 @@ class Profile:
         return period
 
     def Ra(self):
-        return np.abs(self._data - self._data.mean()).sum() / self._data.size
+        return np.abs(self.data - self.data.mean()).sum() / self.data.size
 
     def Rq(self):
-        return np.sqrt(((self._data - self._data.mean()) ** 2).sum() / self._data.size)
+        return np.sqrt(((self.data - self.data.mean()) ** 2).sum() / self.data.size)
 
     def Rp(self):
-        return (self._data - self._data.mean()).max()
+        return (self.data - self.data.mean()).max()
 
     def Rv(self):
-        return np.abs((self._data - self._data.mean()).min())
+        return np.abs((self.data - self.data.mean()).min())
 
     def Rz(self):
         return self.Rp() + self.Rv()
 
     def Rsk(self):
-        return ((self._data - self._data.mean()) ** 3).sum() / self._data.size / self.Rq() ** 3
+        return ((self.data - self.data.mean()) ** 3).sum() / self.data.size / self.Rq() ** 3
 
     def Rku(self):
-        return ((self._data - self._data.mean()) ** 4).sum() / self._data.size / self.Rq() ** 4
+        return ((self.data - self.data.mean()) ** 4).sum() / self.data.size / self.Rq() ** 4
 
     def show(self):
         fig, ax = plt.subplots(figsize=(10, 3))
-        ax.set_xlim(0, self._length_um)
+        ax.set_xlim(0, self.length)
         ax.set_xlabel('x [µm]')
         ax.set_ylabel('z [µm]')
-        ax.plot(np.linspace(0, self._length_um, self._data.size), self._data, c='k', lw=1)
+        ax.plot(np.linspace(0, self.length, self.data.size), self.data, c='k', lw=1)
         plt.show()